win_probability/requirements.txt
aiohttp>=3.9.0
python-dotenv>=1.0.0
colorama>=0.4.6
requests
certifi>=2024.2.2
curl_cffi>=0.7.0
streamlit
